Rename course feedback data to survey execution data and allow for loading multiple at once
