+++ classes/output/course_event_slot_table.php
@@ -16,7 +16,7 @@
 
 namespace block_coursefeedback\output;
 
-use block_coursefeedback\local\course_feedback_data;
+use block_coursefeedback\local\survey_execution_data;
 use block_coursefeedback\local\persistent\eventtype;
 use block_coursefeedback\local\persistent\response_slot;
 use block_coursefeedback\local\persistent\survey_part_execution;
@@ -49,11 +49,14 @@ class course_event_slot_table implements named_templatable, renderable {
     /**
      * Initialize a new instance.
      *
-     * @param course_feedback_data $course_data
+     * @param survey_execution_data $survey_data
+     * @param object $course
      */
     public function __construct(
-        /** @var course_feedback_data $course_data */
-        private readonly course_feedback_data $course_data
+        /** @var survey_execution_data $survey_data */
+        private readonly survey_execution_data $survey_data,
+        /** @var object $course */
+        private readonly object $course,
     ) {
     }
 
@@ -72,13 +75,13 @@ class course_event_slot_table implements named_templatable, renderable {
      */
     private function export_slot(renderer_base $output, response_slot $slot, bool $allow_deletion): array {
         if ($this->enrolledusers === null) {
-            $this->enrolledusers = array_column(enrol_get_course_users($this->course_data->course->id), null, 'id');
+            $this->enrolledusers = array_column(enrol_get_course_users($this->course->id), null, 'id');
         }
 
         $users_editable = new slot_users_editable(
             slot: $slot,
             availableusers: $this->enrolledusers,
-            assignedusers: $this->course_data->users_by_slot_id[$slot->get('id')] ?? [],
+            assignedusers: $this->survey_data->users_by_slot_id[$slot->get('id')] ?? [],
             editable: true,
         );
 
@@ -98,8 +101,8 @@ class course_event_slot_table implements named_templatable, renderable {
      * @return array
      */
     private function export_spe(renderer_base $output, survey_part_execution $survey_part_execution): array {
-        $survey_part = $this->course_data->survey_parts_by_spe_id[$survey_part_execution->get('id')];
-        $slots = $this->course_data->slots_by_spe_id[$survey_part_execution->get('id')] ?? [];
+        $survey_part = $this->survey_data->survey_parts_by_spe_id[$survey_part_execution->get('id')];
+        $slots = $this->survey_data->slots_by_spe_id[$survey_part_execution->get('id')] ?? [];
         $allow_slot_deletion = count($slots) > 1;
         $first_slot = array_shift($slots);
         return [
@@ -159,7 +162,7 @@ class course_event_slot_table implements named_templatable, renderable {
      */
     private function export_event(renderer_base $output, teaching_event $event): array {
         $id = $event->get('id');
-        $spe = $this->course_data->spes_by_event_id[$id] ?? null;
+        $spe = $this->survey_data->spes_by_event_id[$id] ?? null;
         if (!$spe) {
             throw new coding_exception("Event '$id' has no survey part execution.");
         }
@@ -167,7 +170,7 @@ class course_event_slot_table implements named_templatable, renderable {
             'id' => $id,
             'name' => $event->get('name'),
             'type' => [
-                'name' => $this->course_data->types_by_event_id[$id]->get('name'),
+                'name' => $this->survey_data->types_by_event_id[$id]->get('name'),
             ],
             'available_event_types' => $this->export_available_event_types(selectedid: $event->get('eventtypeid')),
             'survey_part_execution' => $this->export_spe($output, $spe),
@@ -179,15 +182,15 @@ class course_event_slot_table implements named_templatable, renderable {
     public function export_for_template(renderer_base $output): array {
         return [
             'course' => [
-                'id' => $this->course_data->course->id,
-                'shortname' => $this->course_data->course->shortname,
+                'id' => $this->course->id,
+                'shortname' => $this->course->shortname,
             ],
             'survey_execution' => [
-                'id' => $this->course_data->survey_execution->get('id'),
+                'id' => $this->survey_data->survey_execution->get('id'),
             ],
             'events' => array_map(
                 fn($event) => self::export_event($output, $event),
-                array_values($this->course_data->events_by_id)
+                array_values($this->survey_data->events_by_id)
             ),
             'available_event_types' => $this->export_available_event_types(),
             'available_survey_parts' => $this->export_available_survey_parts(),
